console_code/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Env vars & logging
python-dotenv==1.0.1
loguru==0.7.2

# Database (pick what you need)
# MongoDB
motor==3.3.2
pymongo==4.6.1

# SQL Databases (Install only if needed)
sqlalchemy==2.0.34
pymysql==1.1.1
psycopg2-binary==2.9.9

# pydantic model
pydantic[email]==2.6.0   # or your current version

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Utilities
requests==2.32.3
pandas==2.2.3




